gazebo: Update 3D simulation with new freefloating
Fix #10

A better information about the model from Oliver:

The thruster map is a 6x1 vector that links a force of 1 N on the thruster
with the corresponding wrench applied to the body frame.
For example, an horizontal thruster located at 10 cm on the right side of
the body (starboard) will push along the X direction (forward) and also
induce a yaw. Its map will be : <1 0 0 0 0 0.1>

this information is redundant with the thruster link, but this link is not
mandatory and is only used for rendering. Most AUVs have fixed thrusters.
It's simpler to directly define the map instead of actual thruster links.
The links may still be used for graphical rendering, or if they are not
fixed (steering thrusters).

- Olivier Kermorgant

Signed-off-by: Patrick José Pereira <patrickelectric@gmail.com>